possiblecity/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

# python modules
oauth2
markdown
textile
pillow

# django third party modules
south
django-haystack
django-generic-flatblocks
django-imagekit
django-taggit
django-markup
django-oembed

# backbeat apps
-e git+git://github.com/dmeehan/django-backbeat-text.git#egg=django-backbeat-text
-e git+git://github.com/dmeehan/django-backbeat-inlines.git#egg=django-backbeat-inlines
-e git+git://github.com/dmeehan/django-backbeat-twittools.git#egg=django-backbeat-twittools
-e git+git://github.com/dmeehan/django-backbeat-images.git#egg=django-backbeat-images